requirements.txt
Flask==2.3.3
requests==2.31.0
google-cloud-firestore==2.21.0
google-cloud-bigquery==3.11.4
gunicorn==21.2.0
